Don't generate credentials secret getters and setters
We don't expect these to be consistent across all providers. Currently _most_
providers require a credentials secret. The AWS provider does not, because it
supports using a different method (i.e. IRSA) that does not require a secret.

In practice we don't use these getters and setters anywhere where we couldn't
just type assert resource.ProviderConfig to the specific type we're dealing
with.

Signed-off-by: Nic Cope <negz@rk0n.org>